fix: Update RuntimeIdentifier from win10-x64 to win-x64 for .NET SDK … (#4)
* fix: Update RuntimeIdentifier from win10-x64 to win-x64 for .NET SDK 10+ compatibility

- Changed RuntimeIdentifier in TPFan.Service.csproj
- Changed RuntimeIdentifiers in TPFan.UWP.csproj
- Updated build.yml workflow publish command
- Updated msix-packaging.yml workflow publish command

Fixes NETSDK1083 error with .NET SDK 10.0.400

* fix: Add missing using directives in TPFan.Shared for System types

- Add using System; in FanStatus.cs for DateTime
- Add using System; and using System.Linq; in FanCurve.cs for DateTime, Linq methods
- Add using System.Threading.Tasks; in IFanServiceContract.cs for Task<T>

Fixes CS0246 errors with .NET SDK 10.0.400

* fix: Explicitly cast to double in Math.Round to resolve ambiguity

FanCurve.cs line 80 - Math.Round(interpolated) was ambiguous between
decimal and double overloads. Cast to double explicitly.

Fixes CS0121 error with .NET SDK 10.0.400

* fix: Enable ImplicitUsings and add explicit using directives for all projects

- Enable ImplicitUsings in TPFan.Shared, TPFan.Service, and TPFan.UWP csproj
- Add explicit using System, System.Threading.Tasks, System.Collections.Generic
  where needed in Service and UWP projects
- Fix namespace references to use fully qualified names

Fixes CS0246 errors for IDisposable, Task, Task<>, Type and related types

* fix: Qualify ApplicationDataContainerCreateDisposition enum reference

UserSettingsService.cs line 17 - Enum was not accessible without
Windows.Storage qualification. Use fully qualified name.

Fixes CS0103 error

* fix: Remove CreateContainer call and fix namespace references in UWP

- UserSettingsService: Use LocalSettings directly (WinUI 3 compatible)
- FanServiceClient: Fix namespace references to use fully qualified names

ApplicationDataContainerCreateDisposition enum doesn't exist in WinUI 3

+++ TPFan.Shared/Models/FanCurve.cs
@@ -1,3 +1,6 @@
+using System;
+using System.Linq;
+
 namespace TPFan.Shared.Models;
 
 /// <summary>
@@ -74,6 +77,6 @@ public record FanCurve
         var tempDiff = temperatureCelsius - lowerPoint.TemperatureCelsius;
 
         var interpolated = lowerPoint.SpeedPercent + (speedRange * tempDiff) / tempRange;
-        return (int)Math.Round(interpolated);
+        return (int)Math.Round((double)interpolated);
     }
 }
